Update README with detailed user instructions
Instructions on how to properly setup the project has been added to the README file. This includes steps on how to fork the project, add Docker Hub credentials, and activate the Renovate. Guidance on how to change the image name and start working on one's own project has also been provided.